Add particles on bomb

+++ particle.cpp
@@ -1,7 +1,7 @@
 #include "particle.h"
 
 Defender::Particle::Particle(Defender::Room& room, std::shared_ptr<Defender::Texture> texture,
-                             Vector2D position, int lifeTime, Vector2D velocity,
+                             Vector2D position, double lifeTime, Vector2D velocity,
                              Vector2D acceleration) : Entity(room, texture)
 {
     this->position = position;
